docstore: reorganize files
- Move things to different files.

- Delete redundant example.
