+++ internal/docstore/docstore_test.go
@@ -16,9 +16,6 @@ package docstore
 
 import (
 	"context"
-	"errors"
-	"net/url"
-	"sync"
 	"testing"
 	"time"
 
@@ -26,114 +23,6 @@ import (
 	"gocloud.dev/gcerrors"
 	"gocloud.dev/internal/docstore/driver"
 )
-
-var (
-	testOpenOnce sync.Once
-	testOpenGot  *url.URL
-)
-
-func TestURLMux(t *testing.T) {
-	ctx := context.Background()
-
-	mux := new(URLMux)
-	fake := &fakeOpener{}
-	mux.RegisterCollection("foo", fake)
-	mux.RegisterCollection("err", fake)
-
-	if diff := cmp.Diff(mux.CollectionSchemes(), []string{"err", "foo"}); diff != "" {
-		t.Errorf("Schemes: %s", diff)
-	}
-	if !mux.ValidCollectionScheme("foo") || !mux.ValidCollectionScheme("err") {
-		t.Errorf("ValidCollectionScheme didn't return true for valid scheme")
-	}
-	if mux.ValidCollectionScheme("foo2") || mux.ValidCollectionScheme("http") {
-		t.Errorf("ValidCollectionScheme didn't return false for invalid scheme")
-	}
-
-	for _, tc := range []struct {
-		name    string
-		url     string
-		wantErr bool
-	}{
-		{
-			name:    "empty URL",
-			wantErr: true,
-		},
-		{
-			name:    "invalid URL",
-			url:     ":foo",
-			wantErr: true,
-		},
-		{
-			name:    "invalid URL no scheme",
-			url:     "foo",
-			wantErr: true,
-		},
-		{
-			name:    "unregistered scheme",
-			url:     "bar://mycollection",
-			wantErr: true,
-		},
-		{
-			name:    "func returns error",
-			url:     "err://mycollection",
-			wantErr: true,
-		},
-		{
-			name: "no query options",
-			url:  "foo://mycollection",
-		},
-		{
-			name: "empty query options",
-			url:  "foo://mycollection?",
-		},
-		{
-			name: "using api scheme prefix",
-			url:  "docstore+foo://bar",
-		},
-		{
-			name: "using api+type scheme prefix",
-			url:  "docstore+collection+foo://bar",
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			_, gotErr := mux.OpenCollection(ctx, tc.url)
-			if (gotErr != nil) != tc.wantErr {
-				t.Fatalf("got err %v, want error %v", gotErr, tc.wantErr)
-			}
-			if gotErr != nil {
-				return
-			}
-			if got := fake.u.String(); got != tc.url {
-				t.Errorf("got %q want %q", got, tc.url)
-			}
-			// Repeat with OpenCollectionURL.
-			parsed, err := url.Parse(tc.url)
-			if err != nil {
-				t.Fatal(err)
-			}
-			_, gotErr = mux.OpenCollectionURL(ctx, parsed)
-			if gotErr != nil {
-				t.Fatalf("got err %v want nil", gotErr)
-			}
-			if got := fake.u.String(); got != tc.url {
-				t.Errorf("got %q want %q", got, tc.url)
-			}
-		})
-	}
-}
-
-type fakeOpener struct {
-	u *url.URL // last url passed to OpenCollectionURL
-}
-
-func (o *fakeOpener) OpenCollectionURL(ctx context.Context, u *url.URL) (*Collection, error) {
-	if u.Scheme == "err" {
-		return nil, errors.New("fail")
-	}
-	o.u = u
-	return nil, nil
-}
 
 func TestToDriverMods(t *testing.T) {
 	for _, test := range []struct {
