+++ internal/docstore/memdocstore/mem_test.go
@@ -19,7 +19,6 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"gocloud.dev/gcerrors"
 	"gocloud.dev/internal/docstore"
 	"gocloud.dev/internal/docstore/driver"
 	"gocloud.dev/internal/docstore/drivertest"
@@ -62,6 +61,7 @@ func TestUpdateEncodesValues(t *testing.T) {
 		t.Fatal(err)
 	}
 	coll := docstore.NewCollection(dc)
+	defer coll.Close()
 	doc := docmap{drivertest.KeyField: "testUpdateEncodes", "a": 1}
 	if err := coll.Put(ctx, doc); err != nil {
 		t.Fatal(err)
@@ -92,6 +92,7 @@ func TestUpdateAtomic(t *testing.T) {
 		t.Fatal(err)
 	}
 	coll := docstore.NewCollection(dc)
+	defer coll.Close()
 	doc := docmap{drivertest.KeyField: "testUpdateAtomic", "a": "A", "b": "B"}
 
 	mods := docstore.Mods{"a": "Y", "b.c": "Z"} // "b" is not a map, so "b.c" is an error
@@ -142,6 +143,7 @@ func TestMissingKeyCreateFailsWithKeyFunc(t *testing.T) {
 		t.Fatal(err)
 	}
 	c := docstore.NewCollection(dc)
+	defer c.Close()
 	err = c.Create(context.Background(), map[string]interface{}{})
 	if err == nil {
 		t.Error("got nil, want error")
@@ -179,28 +181,6 @@ func TestSortDocs(t *testing.T) {
 		sortDocs(got, test.field, test.ascending)
 		if diff := cmp.Diff(got, test.want); diff != "" {
 			t.Errorf("%q, asc=%t:\n%s", test.field, test.ascending, diff)
-		}
-	}
-}
-
-func TestAdd(t *testing.T) {
-	for _, test := range []struct {
-		x, y, want interface{}
-		wantCode   gcerrors.ErrorCode
-	}{
-		{int64(1), int64(2), int64(3), gcerrors.OK},
-		{2.5, 3.5, 6.0, gcerrors.OK},
-		{int64(1), 1.5, 2.5, gcerrors.OK},
-		{-1.5, int64(1), -0.5, gcerrors.OK},
-		{"1", 1.1, nil, gcerrors.InvalidArgument},
-		{int64(1), "1", nil, gcerrors.Internal},
-		{1.2, "1", nil, gcerrors.Internal},
-	} {
-		got, err := add(test.x, test.y)
-		gotCode := gcerrors.Code(err)
-		if got != test.want || gotCode != test.wantCode {
-			t.Errorf("add(%v, %v): got (%v, %s), want (%v, %s)",
-				test.x, test.y, got, gotCode, test.want, test.wantCode)
 		}
 	}
 }
